Move FeedImageDataLoaderWithFallbackComposite to production
